Move declaration merging to a separate file
